🔥 chore: move useAuth directory
